LayoutTests/media/modern-media-controls/start-support/start-support-no-source.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<video style="width: 320px; height: 240px;"></video>
<div id="shadow"></div>
<script type="text/javascript">

description("Testing the <code>StartSupport</code> behavior with no source.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

shouldBeTrue("mediaController.controls.showsStartButton");

shadowRoot.host.remove();
media.remove();

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
