LayoutTests/media/modern-media-controls/start-support/start-support-no-source-expected.txt

Testing the StartSupport behavior with no source.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.showsStartButton is true
PASS successfullyParsed is true

TEST COMPLETE
